Fixing types
